requirements.txt
APScheduler==3.6.3
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
dataclasses==0.6
deprecation==2.1.0
dnspython==2.1.0
gotrue==0.2.0
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
postgrest-py==0.4.0
py==1.10.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.18.0
python-telegram-bot==13.7
pytz==2021.1
realtime-py==0.1.2
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
supabase-py==0.0.2
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
websockets==8.1
